+++ bindings/matlab/autogenerated/+iDynTree/AttitudeQuaternionEKFParameters.m
@@ -7,100 +7,100 @@ classdef AttitudeQuaternionEKFParameters < iDynTreeSwigRef
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1709, self);
+        varargout{1} = iDynTreeMEX(1722, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1710, self, varargin{1});
+        iDynTreeMEX(1723, self, varargin{1});
       end
     end
     function varargout = bias_correlation_time_factor(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1711, self);
+        varargout{1} = iDynTreeMEX(1724, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1712, self, varargin{1});
+        iDynTreeMEX(1725, self, varargin{1});
       end
     end
     function varargout = accelerometer_noise_variance(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1713, self);
+        varargout{1} = iDynTreeMEX(1726, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1714, self, varargin{1});
+        iDynTreeMEX(1727, self, varargin{1});
       end
     end
     function varargout = magnetometer_noise_variance(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1715, self);
+        varargout{1} = iDynTreeMEX(1728, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1716, self, varargin{1});
+        iDynTreeMEX(1729, self, varargin{1});
       end
     end
     function varargout = gyroscope_noise_variance(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1717, self);
+        varargout{1} = iDynTreeMEX(1730, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1718, self, varargin{1});
+        iDynTreeMEX(1731, self, varargin{1});
       end
     end
     function varargout = gyro_bias_noise_variance(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1719, self);
+        varargout{1} = iDynTreeMEX(1732, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1720, self, varargin{1});
+        iDynTreeMEX(1733, self, varargin{1});
       end
     end
     function varargout = initial_orientation_error_variance(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1721, self);
+        varargout{1} = iDynTreeMEX(1734, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1722, self, varargin{1});
+        iDynTreeMEX(1735, self, varargin{1});
       end
     end
     function varargout = initial_ang_vel_error_variance(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1723, self);
+        varargout{1} = iDynTreeMEX(1736, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1724, self, varargin{1});
+        iDynTreeMEX(1737, self, varargin{1});
       end
     end
     function varargout = initial_gyro_bias_error_variance(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1725, self);
+        varargout{1} = iDynTreeMEX(1738, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1726, self, varargin{1});
+        iDynTreeMEX(1739, self, varargin{1});
       end
     end
     function varargout = use_magnetometer_measurements(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1727, self);
+        varargout{1} = iDynTreeMEX(1740, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1728, self, varargin{1});
+        iDynTreeMEX(1741, self, varargin{1});
       end
     end
     function self = AttitudeQuaternionEKFParameters(varargin)
@@ -109,14 +109,14 @@ classdef AttitudeQuaternionEKFParameters < iDynTreeSwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = iDynTreeMEX(1729, varargin{:});
+        tmp = iDynTreeMEX(1742, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.SwigClear();
       end
     end
     function delete(self)
       if self.swigPtr
-        iDynTreeMEX(1730, self);
+        iDynTreeMEX(1743, self);
         self.SwigClear();
       end
     end