+++ bindings/matlab/autogenerated/+iDynTree/BerdySensor.m
@@ -7,37 +7,37 @@ classdef BerdySensor < iDynTreeSwigRef
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1566, self);
+        varargout{1} = iDynTreeMEX(1579, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1567, self, varargin{1});
+        iDynTreeMEX(1580, self, varargin{1});
       end
     end
     function varargout = id(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1568, self);
+        varargout{1} = iDynTreeMEX(1581, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1569, self, varargin{1});
+        iDynTreeMEX(1582, self, varargin{1});
       end
     end
     function varargout = range(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1570, self);
+        varargout{1} = iDynTreeMEX(1583, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1571, self, varargin{1});
+        iDynTreeMEX(1584, self, varargin{1});
       end
     end
     function varargout = eq(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1572, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1585, self, varargin{:});
     end
     function varargout = lt(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1573, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1586, self, varargin{:});
     end
     function self = BerdySensor(varargin)
       if nargin==1 && strcmp(class(varargin{1}),'iDynTreeSwigRef')
@@ -45,14 +45,14 @@ classdef BerdySensor < iDynTreeSwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = iDynTreeMEX(1574, varargin{:});
+        tmp = iDynTreeMEX(1587, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.SwigClear();
       end
     end
     function delete(self)
       if self.swigPtr
-        iDynTreeMEX(1575, self);
+        iDynTreeMEX(1588, self);
         self.SwigClear();
       end
     end