+++ bindings/matlab/autogenerated/+iDynTree/ForwardPosVelAccKinematics.m
@@ -1,3 +1,3 @@
 function varargout = ForwardPosVelAccKinematics(varargin)
-  [varargout{1:nargout}] = iDynTreeMEX(1217, varargin{:});
+  [varargout{1:nargout}] = iDynTreeMEX(1232, varargin{:});
 end