+++ bindings/matlab/autogenerated/+iDynTree/ForwardVelAccKinematics.m
@@ -1,3 +1,3 @@
 function varargout = ForwardVelAccKinematics(varargin)
-  [varargout{1:nargout}] = iDynTreeMEX(1216, varargin{:});
+  [varargout{1:nargout}] = iDynTreeMEX(1231, varargin{:});
 end