+++ bindings/matlab/autogenerated/+iDynTree/GyroscopeSensor.m
@@ -7,55 +7,55 @@ classdef GyroscopeSensor < iDynTree.LinkSensor
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = iDynTreeMEX(1376, varargin{:});
+        tmp = iDynTreeMEX(1391, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.SwigClear();
       end
     end
     function delete(self)
       if self.swigPtr
-        iDynTreeMEX(1377, self);
+        iDynTreeMEX(1392, self);
         self.SwigClear();
       end
     end
     function varargout = setName(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1378, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1393, self, varargin{:});
     end
     function varargout = setLinkSensorTransform(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1379, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1394, self, varargin{:});
     end
     function varargout = setParentLink(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1380, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1395, self, varargin{:});
     end
     function varargout = setParentLinkIndex(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1381, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1396, self, varargin{:});
     end
     function varargout = getName(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1382, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1397, self, varargin{:});
     end
     function varargout = getSensorType(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1383, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1398, self, varargin{:});
     end
     function varargout = getParentLink(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1384, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1399, self, varargin{:});
     end
     function varargout = getParentLinkIndex(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1385, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1400, self, varargin{:});
     end
     function varargout = getLinkSensorTransform(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1386, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1401, self, varargin{:});
     end
     function varargout = isValid(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1387, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1402, self, varargin{:});
     end
     function varargout = clone(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1388, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1403, self, varargin{:});
     end
     function varargout = updateIndices(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1389, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1404, self, varargin{:});
     end
     function varargout = predictMeasurement(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1390, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1405, self, varargin{:});
     end
   end
   methods(Static)