+++ bindings/matlab/autogenerated/+iDynTree/InverseDynamicsInertialParametersRegressor.m
@@ -1,3 +1,3 @@
 function varargout = InverseDynamicsInertialParametersRegressor(varargin)
-  [varargout{1:nargout}] = iDynTreeMEX(1248, varargin{:});
+  [varargout{1:nargout}] = iDynTreeMEX(1263, varargin{:});
 end