+++ bindings/matlab/autogenerated/+iDynTree/createModelWithNormalizedJointNumbering.m
@@ -1,0 +1,3 @@
+function varargout = createModelWithNormalizedJointNumbering(varargin)
+  [varargout{1:nargout}] = iDynTreeMEX(1167, varargin{:});
+end