+++ bindings/matlab/autogenerated/+iDynTree/dynamicsEstimationForwardVelKinematics.m
@@ -1,3 +1,3 @@
 function varargout = dynamicsEstimationForwardVelKinematics(varargin)
-  [varargout{1:nargout}] = iDynTreeMEX(1517, varargin{:});
+  [varargout{1:nargout}] = iDynTreeMEX(1530, varargin{:});
 end