+++ bindings/matlab/autogenerated/+iDynTree/estimateExternalWrenchesBuffers.m
@@ -9,86 +9,76 @@ classdef estimateExternalWrenchesBuffers < iDynTreeSwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = iDynTreeMEX(1496, varargin{:});
+        tmp = iDynTreeMEX(1511, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.SwigClear();
       end
     end
     function varargout = resize(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1497, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1512, self, varargin{:});
     end
     function varargout = getNrOfSubModels(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1498, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1513, self, varargin{:});
     end
     function varargout = getNrOfLinks(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1499, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1514, self, varargin{:});
     end
     function varargout = isConsistent(self,varargin)
-      [varargout{1:nargout}] = iDynTreeMEX(1500, self, varargin{:});
+      [varargout{1:nargout}] = iDynTreeMEX(1515, self, varargin{:});
     end
     function varargout = A(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1501, self);
+        varargout{1} = iDynTreeMEX(1516, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1502, self, varargin{1});
+        iDynTreeMEX(1517, self, varargin{1});
       end
     end
     function varargout = x(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1503, self);
+        varargout{1} = iDynTreeMEX(1518, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1504, self, varargin{1});
+        iDynTreeMEX(1519, self, varargin{1});
       end
     end
     function varargout = b(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1505, self);
+        varargout{1} = iDynTreeMEX(1520, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1506, self, varargin{1});
-      end
-    end
-    function varargout = pinvA(self, varargin)
-      narginchk(1, 2)
-      if nargin==1
-        nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1507, self);
-      else
-        nargoutchk(0, 0)
-        iDynTreeMEX(1508, self, varargin{1});
+        iDynTreeMEX(1521, self, varargin{1});
       end
     end
     function varargout = b_contacts_subtree(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1509, self);
+        varargout{1} = iDynTreeMEX(1522, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1510, self, varargin{1});
+        iDynTreeMEX(1523, self, varargin{1});
       end
     end
     function varargout = subModelBase_H_link(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = iDynTreeMEX(1511, self);
+        varargout{1} = iDynTreeMEX(1524, self);
       else
         nargoutchk(0, 0)
-        iDynTreeMEX(1512, self, varargin{1});
+        iDynTreeMEX(1525, self, varargin{1});
       end
     end
     function delete(self)
       if self.swigPtr
-        iDynTreeMEX(1513, self);
+        iDynTreeMEX(1526, self);
         self.SwigClear();
       end
     end