+++ bindings/matlab/autogenerated/+iDynTree/isDOFBerdyDynamicVariable.m
@@ -1,3 +1,3 @@
 function varargout = isDOFBerdyDynamicVariable(varargin)
-  [varargout{1:nargout}] = iDynTreeMEX(1546, varargin{:});
+  [varargout{1:nargout}] = iDynTreeMEX(1559, varargin{:});
 end