+++ bindings/matlab/autogenerated/+iDynTree/isLinkBerdyDynamicVariable.m
@@ -1,3 +1,3 @@
 function varargout = isLinkBerdyDynamicVariable(varargin)
-  [varargout{1:nargout}] = iDynTreeMEX(1544, varargin{:});
+  [varargout{1:nargout}] = iDynTreeMEX(1557, varargin{:});
 end