+++ bindings/matlab/autogenerated/+iDynTree/sizeOfRotationParametrization.m
@@ -1,3 +1,3 @@
 function varargout = sizeOfRotationParametrization(varargin)
-  [varargout{1:nargout}] = iDynTreeMEX(2039, varargin{:});
+  [varargout{1:nargout}] = iDynTreeMEX(2052, varargin{:});
 end