+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix compilation against PyPy (https://github.com/robotology/idyntree/pull/1018).
 
+### Changed
+- estimateExternalWrenches: Avoid to compute the pseudoinverse to evaluate the external wrenches (https://github.com/robotology/idyntree/pull/1017).
+
 ## [6.1.0] - 2022-08-18
 
 ### Added